move LinkedQueue to utils folder
