+++ example/example/ExecutionQueue.swift
@@ -44,30 +44,3 @@ class ExecutionQueue {
   }
   
 }
-
-// TODO: move to utils folder
-public struct LinkedQueue<T> {
-  
-  fileprivate var list = LinkedList<T>()
-
-  public var isEmpty: Bool {
-    return list.isEmpty
-  }
-  
-  public var count: Int {
-    return list.count
-  }
-  
-  public mutating func add(_ element: T) {
-    list.append(element)
-  }
-  
-  public mutating func remove() -> T? {
-    if isEmpty {
-      return nil
-    } else {
-      return list.removeLast()
-    }
-  }
-  
-}
